// src/types/enums.ts
export enum RowType {
    TEACHER = "teacher",
    STUDENT = "student",
    PARENT = "parent",
    SUBJECT = "subject",
    CLASS = "class",
    LESSON = "lesson",
    EXAM = "exam",
    ASSIGNMENT = "assignment",
    RESULT = "result",
    EVENT = "event",
    ANNOUNCEMENT = 'annoucement'
}

export enum FormType{
    CREATE = 'create',
    UPDATE = 'update',
    DELETE = 'delete'
}

export enum FormFieldType{
    INPUT = 'input',
    SELECT = 'select',
    DATE = 'date'
}